+++ .github/workflows/build.yml
@@ -1,0 +1,22 @@
+name: Build
+
+on:
+  pull_request:
+
+jobs:
+  build-app:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout repo
+        uses: actions/checkout@v4
+
+      - name: Set up JDK 17
+        uses: actions/setup-java@v4
+        with:
+          distribution: 'temurin'
+          java-version: '17'
+          cache: maven
+
+      - name: Build application
+        run: mvn verify --no-transfer-progress
