+++ hw01/src/main/java/ru/otus/hw/Application.java
@@ -1,0 +1,15 @@
+package ru.otus.hw;
+
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+import ru.otus.hw.service.TestRunnerService;
+
+public class Application {
+    public static void main(String[] args) {
+        
+        //Прописать бины в spring-context.xml и создать контекст на его основе
+        ApplicationContext context = new ClassPathXmlApplicationContext("spring-context.xml");
+        var testRunnerService = context.getBean(TestRunnerService.class);
+        testRunnerService.run();
+    }
+}
