+++ hw01/src/main/java/ru/otus/hw/config/AppProperties.java
@@ -1,0 +1,10 @@
+package ru.otus.hw.config;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@AllArgsConstructor
+@Data
+public class AppProperties implements TestFileNameProvider {
+    private String testFileName;
+}
