+++ hw01/src/main/java/ru/otus/hw/config/TestFileNameProvider.java
@@ -1,0 +1,5 @@
+package ru.otus.hw.config;
+
+public interface TestFileNameProvider {
+    String getTestFileName();
+}
