+++ hw01/src/main/java/ru/otus/hw/dao/QuestionDao.java
@@ -1,0 +1,8 @@
+package ru.otus.hw.dao;
+
+import java.util.List;
+import ru.otus.hw.domain.Question;
+
+public interface QuestionDao {
+  List<Question> findAll();
+}
