+++ hw01/src/main/java/ru/otus/hw/dao/dto/AnswerCsvConverter.java
@@ -1,0 +1,13 @@
+package ru.otus.hw.dao.dto;
+
+import com.opencsv.bean.AbstractCsvConverter;
+import ru.otus.hw.domain.Answer;
+
+public class AnswerCsvConverter extends AbstractCsvConverter {
+    
+    @Override
+    public Object convertToRead(String value) {
+        var valueArr = value.split("%");
+        return new Answer(valueArr[0], Boolean.parseBoolean(valueArr[1]));
+    }
+}
