+++ hw01/src/main/java/ru/otus/hw/domain/Answer.java
@@ -1,0 +1,4 @@
+package ru.otus.hw.domain;
+
+public record Answer(String text, boolean isCorrect) {
+}
