+++ hw01/src/main/java/ru/otus/hw/domain/Question.java
@@ -1,0 +1,6 @@
+package ru.otus.hw.domain;
+
+import java.util.List;
+
+public record Question(String text, List<Answer> answers) {
+}
