+++ hw01/src/main/java/ru/otus/hw/exceptions/QuestionReadException.java
@@ -1,0 +1,11 @@
+package ru.otus.hw.exceptions;
+
+public class QuestionReadException extends RuntimeException {
+    public QuestionReadException(String message, Throwable ex) {
+        super(message, ex);
+    }
+    
+    public QuestionReadException(String message) {
+        super(message);
+    }
+}
