+++ hw01/src/main/java/ru/otus/hw/service/IOService.java
@@ -1,0 +1,7 @@
+package ru.otus.hw.service;
+
+public interface IOService {
+    void printLine(String s);
+    
+    void printFormattedLine(String s, Object ...args);
+}
