+++ hw01/src/main/java/ru/otus/hw/service/StreamsIOService.java
@@ -1,0 +1,22 @@
+package ru.otus.hw.service;
+
+import java.io.PrintStream;
+
+public class StreamsIOService implements IOService {
+    private final PrintStream printStream;
+    
+    public StreamsIOService(PrintStream printStream) {
+        
+        this.printStream = printStream;
+    }
+    
+    @Override
+    public void printLine(String s) {
+        printStream.println(s);
+    }
+    
+    @Override
+    public void printFormattedLine(String s, Object... args) {
+        printStream.printf(s + "%n", args);
+    }
+}
