+++ hw01/src/main/java/ru/otus/hw/service/TestRunnerService.java
@@ -1,0 +1,5 @@
+package ru.otus.hw.service;
+
+public interface TestRunnerService {
+    void run();
+}
