+++ hw01/src/main/java/ru/otus/hw/service/TestRunnerServiceImpl.java
@@ -1,0 +1,14 @@
+package ru.otus.hw.service;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class TestRunnerServiceImpl implements TestRunnerService {
+    
+    private final TestService testService;
+    
+    @Override
+    public void run() {
+        testService.executeTest();
+    }
+}
