+++ hw01/src/main/java/ru/otus/hw/service/TestService.java
@@ -1,0 +1,6 @@
+package ru.otus.hw.service;
+
+public interface TestService {
+    
+    void executeTest();
+}
