+++ hw01/src/test/resources/questions_test_file.csv
@@ -1,0 +1,2 @@
+# Empty line for skipping
+What is the capital of France?;Paris%true|London%false|Berlin%false
