--- src/main/java/ru/otus/hw01/config/AppProperties.java
@@ -1,10 +1,0 @@
-package ru.otus.hw01.config;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@AllArgsConstructor
-@Data
-public class AppProperties implements TestFileNameProvider {
-    private String testFileName;
-}
