--- src/main/java/ru/otus/hw01/config/TestFileNameProvider.java
@@ -1,5 +1,0 @@
-package ru.otus.hw01.config;
-
-public interface TestFileNameProvider {
-    String getTestFileName();
-}
