--- src/main/java/ru/otus/hw01/dao/QuestionDao.java
@@ -1,9 +1,0 @@
-package ru.otus.hw01.dao;
-
-import ru.otus.hw01.domain.Question;
-
-import java.util.List;
-
-public interface QuestionDao {
-  List<Question> findAll();
-}
