--- src/main/java/ru/otus/hw01/dao/dto/AnswerCsvConverter.java
@@ -1,13 +1,0 @@
-package ru.otus.hw01.dao.dto;
-
-import com.opencsv.bean.AbstractCsvConverter;
-import ru.otus.hw01.domain.Answer;
-
-public class AnswerCsvConverter extends AbstractCsvConverter {
-    
-    @Override
-    public Object convertToRead(String value) {
-        var valueArr = value.split("%");
-        return new Answer(valueArr[0], Boolean.parseBoolean(valueArr[1]));
-    }
-}
