--- src/main/java/ru/otus/hw01/dao/dto/QuestionDto.java
@@ -1,25 +1,0 @@
-package ru.otus.hw01.dao.dto;
-
-import com.opencsv.bean.CsvBindAndSplitByPosition;
-import com.opencsv.bean.CsvBindByPosition;
-import lombok.Data;
-import ru.otus.hw01.domain.Answer;
-import ru.otus.hw01.domain.Question;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Data
-public class QuestionDto {
-    
-    @CsvBindByPosition(position = 0)
-    private String text;
-    
-    @CsvBindAndSplitByPosition(position = 1, collectionType = ArrayList.class, elementType = Answer.class,
-        converter = AnswerCsvConverter.class, splitOn = "\\|")
-    private List<Answer> answers;
-    
-    public Question toDomainObject() {
-        return new Question(text, answers);
-    }
-}
