--- src/main/java/ru/otus/hw01/domain/Answer.java
@@ -1,4 +1,0 @@
-package ru.otus.hw01.domain;
-
-public record Answer(String text, boolean isCorrect) {
-}
