--- src/main/java/ru/otus/hw01/domain/Question.java
@@ -1,6 +1,0 @@
-package ru.otus.hw01.domain;
-
-import java.util.List;
-
-public record Question(String text, List<Answer> answers) {
-}
