--- src/main/java/ru/otus/hw01/service/IOService.java
@@ -1,7 +1,0 @@
-package ru.otus.hw01.service;
-
-public interface IOService {
-    void printLine(String s);
-    
-    void printFormattedLine(String s, Object ...args);
-}
