--- src/main/java/ru/otus/hw01/service/TestRunnerService.java
@@ -1,5 +1,0 @@
-package ru.otus.hw01.service;
-
-public interface TestRunnerService {
-    void run();
-}
