--- src/main/java/ru/otus/hw01/service/TestRunnerServiceImpl.java
@@ -1,14 +1,0 @@
-package ru.otus.hw01.service;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class TestRunnerServiceImpl implements TestRunnerService {
-    
-    private final TestService testService;
-    
-    @Override
-    public void run() {
-        testService.executeTest();
-    }
-}
