--- src/main/java/ru/otus/hw01/service/TestService.java
@@ -1,6 +1,0 @@
-package ru.otus.hw01.service;
-
-public interface TestService {
-    
-    void executeTest();
-}
