--- src/main/resources/spring-context.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns="http://www.springframework.org/schema/beans"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans
-        https://www.springframework.org/schema/beans/spring-beans-4.3.xsd">
-
-    <bean id="ioService" class="ru.otus.hw01.service.StreamsIOService">
-        <constructor-arg index="0" value="#{T(java.lang.System).out}"/>
-    </bean>
-    <bean id="testRunnerService" class="ru.otus.hw01.service.TestRunnerServiceImpl">
-        <constructor-arg ref="testService"/>
-    </bean>
-    <bean id="testService" class="ru.otus.hw01.service.TestServiceImpl">
-        <constructor-arg ref="ioService"/>
-        <constructor-arg ref="questionDao"/>
-    </bean>
-    <bean id="appProperties" class="ru.otus.hw01.config.AppProperties">
-        <constructor-arg value="#{'questions.csv'}"/>
-    </bean>
-    <bean id="questionDao" class="ru.otus.hw01.dao.CsvQuestionDao">
-        <constructor-arg ref="appProperties"/>
-    </bean>
-
-</beans>
